requirements.txt
annotated-types==0.7.0
anyio==4.4.0
asgiref==3.8.1
cachetools==5.4.0
certifi==2024.7.4
cffi==1.16.0
charset-normalizer==3.3.2
colorama==0.4.6
cryptography==43.0.0
distro==1.9.0
Django==4.2.15
django-allauth==0.63.3
django-widget-tweaks==1.5.0
exceptiongroup==1.2.2
google-api-core==2.19.1
google-api-python-client==2.137.0
google-auth==2.32.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.1
googleapis-common-protos==1.63.2
h11==0.14.0
httpcore==1.0.5
httplib2==0.22.0
httpx==0.27.0
idna==3.7
oauthlib==3.2.2
openai==1.35.12
proto-plus==1.24.0
protobuf==5.27.2
psycopg2==2.9.9
psycopg2-binary==2.9.9
pyasn1==0.6.0
pyasn1-modules==0.4.0
pycparser==2.22
pydantic==2.8.2
pydantic-core==2.20.1
pyparsing==3.1.2
requests==2.32.3
requests-oauthlib==2.0.0
rsa==4.9
ruff==0.5.1
sniffio==1.3.1
sqlparse==0.5.0
tqdm==4.66.4
typing-extensions==4.12.2
tzdata==2024.1
uritemplate==4.1.1
urllib3==2.2.2